Filled config struct

+++ task 3 - basic software architecture/SAMV71x/app/08_UartBase/src/Bsw/Mcal/Com/Uart/Uart_Cfg.h
@@ -59,7 +59,7 @@
   typedef struct UartConfigType
   {
     uint8_t UartNumberOfChannels;
-    uint8_t ClkSrc;
+    uint32_t ClkSrc;
     UartChannelType* UartChannel;
   } UartConfigType;
